plugins/apel/pyproject.toml
[build-system]
requires = ["setuptools==75.2.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auditor_apel_plugin"
version = "0.6.2"
requires-python = ">=3.8"
dependencies = [
	     "python-auditor==0.6.2",
	     "requests==2.32.3",
	     "cryptography==43.0.3",
	     "pyyaml==6.0.2",
	     "pydantic==2.9.2",
]
description = "AUDITOR plugin for sending accounting data to APEL"
readme = "README.md"

[project.optional-dependencies]
style = [
      "black==24.10.0",
      "ruff==0.7.0",
]
tests = [
      "pytest==8.3.3",
      "pytest-cov==5.0.0",
      "mypy==1.13.0",	
      "types-pyyaml==6.0.12.20240917",
      "types-requests==2.32.0.20241016",
]
build = [
      "build==1.2.2.post1",
      "setuptools==75.2.0",
]

[project.scripts]
auditor-apel-publish = "auditor_apel_plugin.publish:main"
auditor-apel-republish = "auditor_apel_plugin.republish:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["*__init__.py", "*publish.py", "*_version.py"]
branch = true

[tool.coverage.lcov]
output = "lcov.info"

[tool.black]
include = '\.pyi?$'

[tool.pytest.ini_options]
addopts = "-s -v --cov --cov-report lcov"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[[tool.mypy.overrides]]
module = "pyauditor"
ignore_missing_imports = true

[tool.ruff.lint]
extend-select = ["I"]